+++ src/pipirik_wars/application/forest/__init__.py
@@ -1,17 +1,25 @@
 """Use-cases леса (Спринт 1.3)."""
 
+from pipirik_wars.application.forest.apply_name_drop import (
+    ApplyForestNameDrop,
+    ForestNameDropApplied,
+)
 from pipirik_wars.application.forest.finish_run import (
     FinishForestRun,
     ForestRunFinished,
 )
+from pipirik_wars.application.forest.notifier import IForestFinishNotifier
 from pipirik_wars.application.forest.start_run import (
     ForestRunStarted,
     StartForestRun,
 )
 
 __all__ = [
+    "ApplyForestNameDrop",
     "FinishForestRun",
+    "ForestNameDropApplied",
     "ForestRunFinished",
     "ForestRunStarted",
+    "IForestFinishNotifier",
     "StartForestRun",
 ]